enabled partial evaluation of Fractional-base radicals
Previously, a factor of a radical base was only removed from the radical if its multiplicity was greater than or equal to the index of the radical. Now, the multiplicity need only be greater than or equal to a divisor of the index. For example, 8^(1/4) now evaluates to 2^(1/4)2^(1/2) rather than staying as is.