integrated NumberList
Exponentiations with Rational bases and Fractional exponents now evaluate to NumberLists consisting of the basic evaluation of the exponentiation repeatedly multiplied by the relevant root of unity, expressed in exponential form. This should be refined so that if the exponentiation remains unsimplified, no list is generated. It's possible that intermediate cases also need to be considered separately. Moreover, the handling of roots of unity should be refined so they take complex number rather than exponential form at least in trivial cases.