generalized cos()
In theory, cos() now evaluates to expressions in terms of radicals for any fractional argument whose denominator is 1, 3, 5, 15 or 17 times a power of 2. However, in practice exponentiate() still can't resolve all roots for radicals of index greater than 4. This is because exponentiate() makes a cos() call whose argument has a denominator that is as much as 4 times the index of the radical - for an index of 5, the argument is 9/20 - and for such large denominators, the fact that the program doesn't handle radical denesting means that the evaluation for cos() it calculates becomes extremely complicated. For the argument 9/20, a divide by 0 exception throws several iterations into the loop on line 1118, by which point the value of cosine is so complicated that it would be very difficult to heck whether it's due to a logic error on my part or perhaps some kind of overflow. Radical denesting could probably solve the problem, and is an eventual goal since nested radicals can occur directly in user input and should certainly be simplified in that case. However, adding a sin() function would be easier in the short term, and quite likely more efficient in this respect in the long term. I might be able to get away without also adding a Sine class and thereby avoid the trouble of simulating all the ways sines and cosines can interact and combine.