Number class hierarchy restructure
GaussianInteger makes more sense as a derivation of ComplexNumber than beside Integer. Since this change puts Integer and GaussianInteger at opposite ends of the Number class hierarchy, it was necessary to change WholeNumber from an abstract class to an interface, and move the implementation it contained elsewhere. This mostly entailed inlining the implementation in the places where it was called.