refinement to calculating products of surds
Previously, a product of surds would be left as is as if neither the bases nor the indices matched. Now, the surds are consolidated into a single surd or greater or equal index, which is then passed through exponentiate(), guaranteeing that the lagest index in the resulting expression is as low as possible. For example, 6^(1/3)*3^(1/6)=2^(1/3)*3^(1/2). Exponentiation was dived int Surd and Transcendental classes to facilitate this change. The algorithm for taking roots of Gaussian integers was also removed, because that functionality will be more appropriately handled by radical denesting. Note that at the moment, in the case of roots of negative integers, the negative is left inside the surd.