reinstated handling of roots of negative integers
Also removed the feature of returning all roots, including the NumberList class. The feature assumed that when the user enters an nth root, the notation stands for all n of the complex roots, but in general the output requires being able to represent nth roots that stand for only a single value. I preferred leaving out the feature to coming up with a notational distinction to mark how many roots are intended in order continue treating nth root notation in this nonstandard way. It would probably have been messy to determine how such a system would have interacted with, for example, radical denesting, and it was also unclear whether, for example, 4^(1/2)*4^(1/2) should equal {-2,2} while (4^(1/2))^2 should equal 2 only.