replaced Sum.reciprocal() minimal polynomial handling
I believe MultivariatePolynomial.getMinimalPoynomial() can yield a polynomial that is annulling but not minimal if the numbers represented by this's variables are linearly dependent over the rationals; a case that would ideally be avoided by the way operations on Number objects are implemented, but that isn't currently. More tentatively, the matrix triangularization step might fail altogether in this case.