implemented Sum.reciprocal()
Could fail for Sums whose terms are linearly dependent, because GetMinimalPolynomial() might give an annulling polynomial that isn't minimal in that case. The Number operator implementations are also not guaranteed to succeed in simplifying the product of a Sum's conjugates to an integer. If the product evaluates to another Sum, the result will be a stack overflow. Apparently the integer the multiplication should work out to can alternatively be calculated using some kind of determinant; I should look into using that method instead of conjugate multiplication, because unlike the latter, it's probably foolproof.