switched to sparse MultivariatePolynomial representation
This yields annullingPolynomialCoefficients objects with an equal or lesser count than with the dense representation in MultivariablePolynomial.getMinimalPolynomial(). Since the polynomial specified by annullingPolynomialCoefficients has to be factored, reducing its count should appreciably calculation speed in those cases. Unfortunately, the improvement isn't enough for the input 1/(1+2^(1/3)) to calculate with tolerable speed.