streamlined conjugate candidate testing
This avoids the polynomial factorization bottleneck for the input 1/(1+2^(1/3)) (which will still arise for higher-degree inputs). Now the reason 1/(1+2^(1/3)) fails to compute is that the Sum multiplication procedure fails to simplify the product of the denominator's conjugates to an Integer. This problem can be avoided by using the linear algebra approach to calculate the field norm instead of multiplying the conjugates.