disallowed Polynomials with Fraction coefficients
Polynomial.getGCD() now uses the subresultant algorithm, which should be a performance improvement, but the primary motivation for the change is to make the introduction of Polynomial arithmetic over finite fields, for use in Polynomial factorization, neater.

The input 1/(1+2^(1/3)) now crashes during the what used to be infinite loop the program falls into when it fails to simplify the norm of the denominator to an Integer. The crash happens upon trying to divide the arguments of Polynomial.getGCD() by their contents once the function is called on Polynomials with very large coefficients - an operation that should always have yielded Integer coefficients, but that wasn't enforced by a cast until now. The operation is so elementary that the only candidate for the source of the bug would seem to be a problem with large Integer arithmetic. While I tested arithmetic between a small integer and a large integer, I may have neglected arithmetic between pairs of large integers.