bullet-proofed minimal polynomials
A thorough overhaul. In order to get error-free minimal polynomials, it was necessary to be able to test whether a set of Numbers is linearly independent over the rationals, which in turn requires determining whether any of them is a polynomial over the rationals in terms of the others. To this end, NestedPolynomial was added to represent number field elements, and PolynomialIndependentSum for representing a sum as a polynomial in terms of a number with a known minimal polynomial. This infrastructure comes with the benefit of ensuring that the terms of all sums in the output will be linearly independent. Means of numerical approximation, needed for the number field calculations and for correct selection of complex roots when multiplying Surds, were also added. None of this has been debugged yet.