added a Float class for numerical approximation
Each approximation algorithm still uses Rationals internally to simplify the handling of rounding error, but is then rounded outward to a Float so that when one algorithm is applied to the result of another, the denominator explosion doesn't propagate. Since the algorithm used for pi is designed with producing base 16 digits in mind, it may be possible and desirable to alter the Pi class to take advantage of this to yield Float approximations natively.