ModdedPolynomial.GetFactorsOfSquarefree() optimization
Over the course of processing the input 1/(1+2^(1/3)), currently a factorization of the IntegerPolynomial -2197+23898x^3-18438x^6+32560x^9-2112x^12+768x^15+512x^18 is required. There was a bad bottleneck in the exponentiation in Cantor-Zassenhaus split step of ModdedPolynomial.GetFactorsOfSquarefree(), now fixed. There is still a bad bottleneck later, in the Hensel lift step of IntegerPolynomial.GetFactors().