partially fixed RationalPolynomial.EstimateEvaluation
The method currently finds accurate enough bounds on the real and imaginary parts of argument that the real and imaginary parts of the evaluation of *this at those bounds are within errorIntervalSize of each other, and returns them. However, if *this is not monotonic between the bounds on argument, it isn't guaranteed that the exact value of the evaluation is between the bounds returned. Something involving checking the derivative of *this for roots between the bounds on the argument is probably necessary to address this possibility.